[services][scaffold] python script to scaffold a service boilerplate
